+++ Client/src/views/VehicleDetailsView/VehicleDetailsView.jsx
@@ -2,15 +2,16 @@ import React from "react";
 import Navbar from "../../components/Navbar/Navbar";
 import Footer from "../../components/Footer/Footer";
 import Dash from "../../components/dashboard/Dash";
-import ViewVehicles from "../../components/Vehicles/Vehicles";
-const VehiclesView = () => {
+import VehicleDetails from '../../components/VehicleDetails/vehicleDetails'
+
+const VehicleDetailsView = () => {
     return (
         <div>
             <div><Navbar /> </div>
 
                 <div><Dash /></div>
             <div>
-                <div><ViewVehicles /></div>
+                <div><VehicleDetails/></div>
             </div>
 
             <div><Footer /></div>
@@ -18,4 +19,4 @@ const VehiclesView = () => {
     )
 
 }
-export default VehiclesView;
+export default VehicleDetailsView;
